requirements.txt
Django
kafka-python
tweepy==3.9.0
psycopg2
geopy